# app/public/docs/api.yaml
openapi: "3.0.3"
info:
  title: MagicWebb API
  version: "1.0.0"
  description: |
    MagicWebb is a decentralized NFT marketplace on Flare Coston2.
    All `/api/v1/*` routes are rate-limited (60 req/min per IP).
    Routes marked 🔒 require a valid SIWE JWT (Bearer token or `mw_s_*` cookie).
servers:
  - url: https://magicwebb.fly.dev
    description: Production (Flare Coston2 testnet)
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/auctions:
    get:
      summary: List auctions
      description: |
        Filterable by collection, seller and status. `ended` is the derived
        state "active but past ends_at" (awaiting settlement); `all` and an
        empty status apply no filter. Any other value is a 400.
      tags: [Auctions]
      parameters:
        - name: collection
          in: query
          schema: { type: string }
        - name: seller
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [all, active, ended, settled, cancelled], example: "active" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Array of auction rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuctionRow"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auctions/{id}:
    get:
      summary: Get single auction
      tags: [Auctions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 42 }
      responses:
        "200":
          description: Auction row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionRow"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auctions/{id}/bids:
    get:
      summary: Get bids for auction
      tags: [Auctions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Array of bids (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BidRow"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/collections:
    get:
      summary: List tracked collections
      tags: [Collections]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Array of collection rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionRow"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/collections/{address}:
    get:
      summary: Get collection detail with stats
      description: |
        Returns collection metadata + floor price, 24h volume, listed count and
        `verified_reason` (why the badge is what it is, for tooltips).
        Unknown or malformed addresses are 404.
      tags: [Collections]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string, example: "0x..." }
      responses:
        "200":
          description: Collection detail with computed stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/collections/{address}/traits:
    get:
      summary: List trait values for collection
      description: Returns trait_type → [value, ...] map for filter building.
      tags: [Collections]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Trait map
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items: { type: string }
              example:
                Background: ["Red", "Blue", "Gold"]
                Fur: ["Striped", "Spotted"]
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/collections/{address}/tokens:
    get:
      summary: Page through a collection's indexed tokens
      description: |
        Indexed tokens of a collection with current owner, listed flag (and
        price when listed) and last sale, ordered by token id so pages are
        stable. The parent collection's card facts are included once.
        Unknown collections are 404.
      tags: [Collections]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string, example: "0x..." }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 48 }
      responses:
        "200":
          description: One page of tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionTokensPage"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/trending:
    get:
      summary: Trending collections
      tags: [Collections]
      parameters:
        - name: window
          in: query
          schema: { type: string, default: "24h", example: "7d" }
          description: Time window (e.g. 24h, 7d, 30d)
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Array of trending scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendingScore"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Media ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "SIWE JWT via Authorization: Bearer <token>"
    CookieAuth:
      type: apiKey
      in: cookie
      name: mw_s_
      description: "HttpOnly session cookie with addr prefix (e.g. mw_s_0xabcd...)"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "internal error" }
  schemas:
...
    AuctionRow:
      type: object
      properties:
        auction_id:          { type: integer }
        collection:          { type: string }
        token_id:            { type: string }
        seller:              { type: string }
        standard:            { type: string, enum: [erc721, erc1155] }
        reserve_price_wei:   { type: string }
        highest_bid_wei:     { type: string }
        highest_bidder:      { type: string }
        min_increment_bps:   { type: integer }
        starts_at:           { type: string, format: date-time }
        ends_at:             { type: string, format: date-time }
        status:              { type: string, enum: [active, settled, cancelled] }
        create_tx:           { type: string }
        name:                { type: string }
        image_uri:           { type: string }
        collection_verified:   { type: boolean, description: "Collection passed the standard + metadata verifier" }
        collection_creator:    { type: string, description: "Collection ERC-173 owner(); empty when unresolved. Verified + creator = Authentic badge" }
        collection_name:       { type: string }
        collection_tracked:    { type: boolean, description: "True when the indexer tracks the collection (distinguishes unknown contract from known-but-unverified)" }
    BidRow:
      type: object
      properties:
        bidder:    { type: string }
        amount_wei: { type: string }
        tx_hash:   { type: string }
        placed_at: { type: string, format: date-time }

    # ── Offers ────────────────────────────────────────────────────────
...
    CollectionRow:
      type: object
      properties:
        address:      { type: string }
        name:         { type: string }
        symbol:       { type: string }
        standard:     { type: string, enum: [erc721, erc1155] }
        deploy_block: { type: integer }
        verified:     { type: boolean }
        creator_addr: { type: string, description: "ERC-173 owner(); empty when unresolved" }
    VerifiedReason:
      type: object
      description: Why the collection badge is what it is (tooltip breakdown).
      properties:
        standard_ok:   { type: boolean, description: "ERC-165 says standard NFT contract" }
        metadata_ok:   { type: boolean, description: "At least one token has resolved metadata" }
        creator_known: { type: boolean, description: "ERC-173 owner() resolved" }
    CollectionDetail:
      type: object
      allOf:
        - $ref: "#/components/schemas/CollectionRow"
        - type: object
          properties:
            floor_price_wei:  { type: string }
            volume_24h_wei:   { type: string }
            listed_count:     { type: integer }
            verified_reason:
              $ref: "#/components/schemas/VerifiedReason"
    CollectionTokenRow:
      type: object
      properties:
        token_id:      { type: string }
        owner:         { type: string, description: "Current holder (largest holder for ERC-1155)" }
        name:          { type: string }
        image:         { type: string }
        listed:        { type: boolean }
        price_wei:     { type: string, description: "Cheapest active listing; omitted when not listed" }
        last_sale_wei: { type: string, description: "Most recent sale; omitted when never sold" }
    CollectionTokensPage:
      type: object
      properties:
        collection:
          $ref: "#/components/schemas/CollectionRow"
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/CollectionTokenRow"
        page:  { type: integer }
        limit: { type: integer }
        total: { type: integer, description: "Distinct held tokens in the collection" }
...
    TrendingScore:
      type: object
      properties:
        collection: { type: string }
        window:     { type: string }
        score:      { type: number, format: double }
        views:      { type: integer }
        bids:       { type: integer }
        volume_wei: { type: string, description: "Volume in wei" }
        collection_name:     { type: string }
        standard:            { type: string, enum: [erc721, erc1155, ""] }
        collection_verified: { type: boolean }
        collection_creator:  { type: string }
        collection_tracked:  { type: boolean }

    # ── Wallet ────────────────────────────────────────────────────────
...