# backend/internal/api/auctions.go
package api

import (
	"strconv"

	"github.com/gofiber/fiber/v2"
	"github.com/rs/zerolog/log"

	"github.com/OfficialA1manac/MagicWebb/backend/internal/db"
)

// AuctionsService handles auction-related API operations.
type AuctionsService struct {
	q *db.Q
}

// NewAuctionsService creates an AuctionsService.
func NewAuctionsService(q *db.Q) *AuctionsService {
	return &AuctionsService{q: q}
}

// RegisterRoutes registers all auction-related routes under the given router group.
func (s *AuctionsService) RegisterRoutes(api fiber.Router) {
	api.Get("/auctions", s.handleList)
	api.Get("/auctions/:id", s.handleGet)
	api.Get("/auctions/:id/bids", s.handleBids)
}

func (s *AuctionsService) handleList(c *fiber.Ctx) error {
	f := db.AuctionsFilter{
		Collection: c.Query("collection"),
		// token_id narrows to ONE token. Without it the token page had to pull
		// the first 50 auctions in a collection and find its own token client
		// side, which silently found nothing once a collection had more than
		// 50 auctions ahead of it in the ends_at ordering.
		TokenID: c.Query("token_id"),
		Seller:  c.Query("seller"),
		Status:  c.Query("status"),
	}
	// status is validated here rather than in SQL: the column is a Postgres
	// enum, so an unknown value ("all" — which the auctions page sends for
	// its default tab) used to fail the cast and surface as a 500.
	switch f.Status {
	case "", "all", "active", "ended", "settled", "cancelled":
	default:
		return writeErr(c, fiber.StatusBadRequest, "status must be one of all, active, ended, settled, cancelled")
	}
	if lim := c.Query("limit"); lim != "" {
		if n, err := strconv.Atoi(lim); err == nil {
			if n < 1 {
				n = 1
			} else if n > 200 {
				n = 200
			}
			f.Limit = n
		}
	}
	if mp := c.Query("min_price"); mp != "" {
		if !isValidWeiStr(mp) {
			return writeErr(c, fiber.StatusBadRequest, "min_price must be a non-negative integer wei value")
		}
		f.MinPriceWei = mp
	}
	if mp := c.Query("max_price"); mp != "" {
		if !isValidWeiStr(mp) {
			return writeErr(c, fiber.StatusBadRequest, "max_price must be a non-negative integer wei value")
		}
		f.MaxPriceWei = mp
	}
	rows, err := s.q.ListAuctions(c.Context(), f)
	if err != nil {
		return writeErr(c, fiber.StatusInternalServerError, "internal error")
	}
	if rows == nil {
		rows = []db.AuctionRow{}
	}
	if len(rows) == 0 {
		log.Debug().
			Str("collection", f.Collection).
			Str("seller", f.Seller).
			Str("status", f.Status).
			Msg("auctions: query returned zero results — no auctions match the filter criteria")
	}
	return c.JSON(rows)
}

func (s *AuctionsService) handleGet(c *fiber.Ctx) error {
	id, err := strconv.ParseInt(c.Params("id"), 10, 64)
	if err != nil {
		return writeErr(c, fiber.StatusBadRequest, "invalid auction id")
	}
	row, err := s.q.GetAuction(c.Context(), id)
	if err != nil {
		if isNotFound(err) {
			return writeErr(c, fiber.StatusNotFound, "auction not found")
		}
		return writeErr(c, fiber.StatusInternalServerError, "internal error")
	}
	return c.JSON(row)
}

func (s *AuctionsService) handleBids(c *fiber.Ctx) error {
	id, err := strconv.ParseInt(c.Params("id"), 10, 64)
	if err != nil {
		return writeErr(c, fiber.StatusBadRequest, "invalid auction id")
	}
	rows, err := s.q.GetBidsForAuction(c.Context(), id)
	if err != nil {
		return writeErr(c, fiber.StatusInternalServerError, "internal error")
	}
	if rows == nil {
		rows = []db.BidRow{}
	}
	return c.JSON(rows)
}